% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.explor.r.R
\name{label.explor.r}
\alias{label.explor.r}
\title{SPSS Label Interface (label.explor.R)}
\usage{
label.explor.r(myDF, caseid, mycutoff = 5, objections = NULL)
}
\arguments{
\item{myDF}{Dataframe}

\item{caseid}{String name of column of case ids}

\item{mycutoff}{Optional. Change how many levels a factor can have}
}
\value{
Dataframe with factors, added attributes. Ready for use in stats.explor.r
}
\description{
Interface for working with dataframes to prepare for various
analyses using stats.explor.r. Creates labels, can remove
erroneous variables, makes factors, etc.
}
\examples{
label.explor.r()
}
\keyword{Explore}
