% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.explor.r.R
\name{make.me.shiny}
\alias{make.me.shiny}
\title{Write temp table Server interface}
\usage{
make.me.shiny(myvar, myDF.name, myLabels, myDF2, mycutoff)
}
\arguments{
\item{myvar}{string variable name}

\item{tempnum}{Number for temp}

\item{myDFname}{string data frame name}
}
\value{
List of files, first being UI, second being server
}
\description{
For label.explor.r. Outputs as mytable1, mytable2 from tempnum Render table and add to list
}
\examples{
make.me.shiny()
}
\keyword{Explore}
