% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.explor.r.R
\name{myfinaltable}
\alias{myfinaltable}
\title{Write table of labels}
\usage{
myfinaltable(myxnum, myDF)
}
\arguments{
\item{myxnum}{Column number of variable of interest}

\item{myDF}{Dataframe}
}
\value{
Table
}
\description{
For label.explor.r
}
\examples{
myfinaltable()
}
\keyword{Explore}
