% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.explor.r.R
\name{mytempfactor.ui.server}
\alias{mytempfactor.ui.server}
\title{Write 'em both!}
\usage{
mytempfactor.ui.server(tempnum, myvar, myLabels, myDF, myDFname, cutoff,
  should.comma)
}
\arguments{
\item{tempnum}{Number for temp}

\item{myvar}{List of variable names}

\item{myLabels}{List of Label Names}

\item{myDF}{Actual DF}

\item{myDFname}{Name of DF}

\item{cutoff}{Number of variables no longer a factor}

\item{should.comma}{needs a comma or is last?}
}
\value{
File
}
\description{
For label.explor.r. Outputs as mytable1, mytable2 from tempnum
}
\examples{
mytempfactor.ui.server()
}
\keyword{Explore}
