% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{print.multinom}
\alias{print.multinom}
\title{Print method for multinom tables}
\usage{
\method{print}{multinom}(my.print)
}
\arguments{
\item{my.print}{My print}
}
\value{
printed table
}
\description{
Print method for multinom tables
}
