% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.table.R
\name{print.quick.table}
\alias{print.quick.table}
\title{Print Method for quick.table}
\usage{
\method{print}{quick.table}(q.tab, type = 1)
}
\arguments{
\item{q.tab}{quick.table object}

\item{type}{Type of print: 1 = to viewer, 2 = to console}
}
\value{
List of lists with 3 pieces: null model, pre models, and full models
}
\description{
print function for class "quick.table"
}
\keyword{Explore}
