% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.attribs}
\alias{quick.attribs}
\title{Quickly copy attributes from original data frame}
\usage{
quick.attribs(my.new.df, my.orig.df)
}
\arguments{
\item{my.new.df}{New data frame, normally subset of original}

\item{my.orig.df}{Original data frame}
}
\value{
New data frame with all relevant attributes from original df copied over
}
\description{
When you make a subset of data frames and do some other operations,
R unfortunately loses some of the attributes. This will help get them
back for use in label.explor.r
}
