% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.eq.check}
\alias{quick.eq.check}
\title{Check equality}
\usage{
quick.eq.check(val.1, val.2)
}
\arguments{
\item{my.model}{Model to be used}
}
\value{
List of lists with 3 pieces: null model, pre models, and full models
}
\description{
Helper function to make list of null model, pre models,
and full models for calculating type II SS (and eventually
type III SS).
}
\keyword{Explore}
