% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.formula}
\alias{quick.formula}
\title{Formula Dissecter}
\usage{
quick.formula(my.model, my.envir)
}
\arguments{
\item{my.model}{Model to be used}
}
\value{
List of lists with 3 pieces: null model, pre models, and full models
}
\description{
Helper function to make list of null model, pre models,
and full models for calculating type II SS (and eventually
type III SS).
}
\keyword{Explore}
