% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.plots.R
\name{quick.grab}
\alias{quick.grab}
\title{Saves plot for later use}
\usage{
quick.grab()
}
\value{
Saved Plot
}
\description{
Turns whatever plot or graph into a plot and saves
the object.
}
\examples{
my.plot=quick.grab()

}
\keyword{Explore}
