% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.label}
\alias{quick.label}
\title{View SPSS Label}
\usage{
quick.label(variable)
}
\arguments{
\item{variable}{Variable from data set}
}
\value{
String of label
}
\description{
Helper function to see SPSS label
}
\examples{
quick.lavaan(myfit)
}
\keyword{Explore}
