% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.lavaan}
\alias{quick.lavaan}
\title{PixieDust Tables for Lavaan}
\usage{
quick.lavaan(myfit)
}
\arguments{
\item{myfit}{Fit from Lavaan package}
}
\description{
Places html tables in viewer for different areas of Lavaan output
for SEM. Based on NIU class 11/2017
}
\examples{
quick.lavaan(myfit)
}
\keyword{Explore}
