% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.m.test}
\alias{quick.m.test}
\title{Test statistics}
\usage{
quick.m.test(my.matrix, test.stat)
}
\arguments{
\item{myDF}{Dataframe}

\item{partial}{partial search as string}
}
\description{
Helper function to quickly get list of column names
based on partial piece.
}
\examples{
quick.search(myDF,"CAS")
}
\keyword{Explore}
