% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.MICE}
\alias{quick.MICE}
\title{Finish imputation for complex data}
\usage{
quick.MICE(imputed.df, my.new.df)
}
\arguments{
\item{imputed.df}{Mids object from MICE}

\item{my.new.df}{Data frame that has been through label.explor.r}
}
\value{
Data frame with imputed values averaged and replaced.
}
\description{
MICE creates a mids file. This does not always work with R analyses, such as for ordinal.
Therefore, this "averages" the responses of each run for each variable and places them back
in the data frame. Works with factors by taking label information from original data frame,
turning the factors to numbers, and finding the average that way.
}
