% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.multinom.survey}
\alias{quick.multinom.survey}
\title{Survey designs for multinomial logistic functions.}
\usage{
quick.multinom.survey(my.formula, design, my.df, type = "bootstrap",
  replicates = 10)
}
\arguments{
\item{my.formula}{Formula to be used}

\item{design}{Design to be used}

\item{my.df}{Dataframe}

\item{type}{Type of repetition function. see svrepdesign for more info (default: bootstrap)}

\item{replicates}{Number of replications to use (default: 10)}
}
\value{
List of: table of beta, z, probability; ANODE table; omnibus test; Pseudo R-2 values
}
\description{
The "survey" package does not directly compute designs with multinomial functions.
Instead, replicate weights must be computed and applied tediously to every function.
This wrapper function will compute the major pieces of information using your design.
}
