% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.part.cont}
\alias{quick.part.cont}
\title{Partial contrasts for ANOVA and MANOVA tables}
\usage{
quick.part.cont(my.nested.table, adjustment = "bonferroni", latent.cont = F,
  abbrev.length = 30)
}
\arguments{
\item{my.nested.table}{Table of models and SSCP matrices from \code{\link[=quick.reg]{quick.reg()}}}

\item{adjustment}{P-value adjustment sent to p.adjust (default = "bonferroni").}

\item{latent.cont}{Should latent contrasts also be computed? (default = F).}

\item{abbrev.length}{Integer telling how long of a label is too long. Longer than this and will be condensed (default=15)}
}
\value{
Nested table with added rows and information (Unfortunately as character at the moment)
}
\description{
Adds Contrasts and Latent Contrasts to nested table
for creating overall MANOVA tables.
}
\details{
This package takes the SSCP matrices from \code{\link[=quick.reg]{quick.reg()}} and
calculates partial contrasts directly. Also can create
latent contrasts.
}
\examples{
quick.contrast(my.nested.table)
}
\keyword{Explore}
