% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.pooled.multinom}
\alias{quick.pooled.multinom}
\title{Preform weighted multnomial regression with imputed data frames}
\usage{
quick.pooled.multinom(my.formula, weights, imputed.df, include.null = F,
  round = T)
}
\arguments{
\item{my.formula}{Formula to test. Cannot test null model, but can have it returned with any other model}

\item{weights}{Weights to use}

\item{imputed.df}{Imputed df from MICE}

\item{include.null}{Include the null model as well?}

\item{round}{If want values rounded. Default: T}
}
\value{
object of class "pooled", "multinom"
}
\description{
This will perform multinomial logistic regression with the option of including weights
using an imputed data frame from the MICE package. Returns ANODE table, z-table, deviance,
partial R2 statistics, and the analyses. Can return null model as well. Can be combined with
quick.table.multi.model for omnibus table of multiple models.
}
