% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.reg.R
\name{quick.reg.default}
\alias{quick.reg.default}
\title{ANOVA and ANODE tables in HTML}
\usage{
\method{quick.reg}{default}(my.model, my.factor = NULL, show.contrasts = F,
  adjustment = "bonferroni", show.intercepts = T, show.int.change = F,
  part.eta = F, VIF = F, marginality = T, do.return = T,
  abbrev.length = ab.len, show.footer = T, SS.type = 2,
  myDF = my.found.df, type = my.reg.type)
}
\arguments{
\item{my.model}{Model to be tested. Currently supported are lm, glm, clm, manova.}

\item{my.factor}{If there are any factors, list them here.}

\item{show.contrasts}{If there are factors with df>1, show partial contrasts? (default = F). For lm and glm, this is done using Phia, for manova, it is calculated from SSCP matrices.}

\item{adjustment}{Adjustment type to be performed on contrats. Uses p.adjust. (default = "bonferroni")}

\item{show.intercepts}{Show information regarding threshold values for binomial or ordinal regression? (default = T).}

\item{show.int.change}{Show the change in deviance for the intercept? (default = F). This is a diagnostic as to whether your link function is appropriate and/or if you can generalize to the null model.}

\item{part.eta}{Show partial eta square for lm? (default = F).}

\item{VIF}{Show variable inflation factor? (default = F).}

\item{marginality}{Should SS or deviance for the intercept be neglected? (default = T) If set to F, this is a diagnostic tool to see if the effects of the treatment are at least similar in magnitude to the effects of the intercepts.}

\item{do.return}{Return the quick.table or just display it? (default=T)}

\item{abbrev.length}{Abbreviation length for rownames. Default is 30 for clm, 15 otherwise.}

\item{show.footer}{Show footer including missingness, method, and link and adjustment information if appropriate (default = T)}

\item{SS.type}{Type of sums of squares (or deviance) to report. Currently, only type II is reported. (default = 2)}

\item{myDF}{Backup in case can't figure out data frame.}

\item{type}{Backup in case can't figure out type.}
}
\value{
Either quick.table or invisible()
}
\description{
quick.reg method for "lm", "glm", and "clm" classes
}
\examples{
quick.reg(wine.ord)
}
\keyword{Explore}
