% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.reg.R
\name{quick.reg.manova}
\alias{quick.reg.manova}
\title{MANOVA tables in HTML}
\usage{
\method{quick.reg}{manova}(my.model, my.factor = NULL, test.stat = "Pillai",
  show.contrasts = F, adjustment = "bonferroni", show.y.contrasts = F,
  show.latent = F, show.intercepts = F, real.names = T, part.eta = F,
  VIF = F, marginality = T, do.return = T, abbrev.length = ab.len,
  show.footer = T, SS.type = 2, myDF = my.found.df, type = my.reg.type)
}
\arguments{
\item{my.model}{Model to be tested. Currently supported are lm, glm, clm, manova.}

\item{my.factor}{If there are any factors, list them here.}

\item{test.stat}{MANOVA test statistic to use. (default = "Pillai"). Wilks, Pillai, Roy supported.}

\item{show.contrasts}{If there are factors with df>1, show partial contrasts? (default = F). For lm and glm, this is done using Phia, for manova, it is calculated from SSCP matrices.}

\item{adjustment}{Adjustment type to be performed on contrats. Uses p.adjust. (default = "bonferroni")}

\item{show.y.contrasts}{Show how variables load on each variable? (default = F). NOTE: This is NOT the same as latent variables what are latent ANOVAs.}

\item{show.latent}{Show latent variables? (default = F). This is equivalent to running an ANOVA for each independent variable while making the dependent variables as uncorrelated as possible.}

\item{show.intercepts}{Show the individual intercepts? (default = F).}

\item{real.names}{Use real names for intercepts in contrasts or use intercept number? (default = T). This does not change intercept names if show.intercepts is on.}

\item{part.eta}{NOT YET SUPPORTED}

\item{VIF}{NOT YET SUPPORTED}

\item{marginality}{Should SS or deviance for the intercept be neglected? (default = T) If set to F, this is a diagnostic tool to see if the effects of the treatment are at least similar in magnitude to the effects of the intercepts.}

\item{do.return}{Return the quick.table or just display it? (default=T)}

\item{abbrev.length}{Abbreviation length for rownames. Default is 30 for clm, 15 otherwise.}

\item{show.footer}{Show footer including missingness, method, and link and adjustment information if appropriate (default = T)}

\item{SS.type}{Type of sums of squares (or deviance) to report. Currently, only type II is reported. (default = 2)}

\item{myDF}{Backup in case can't figure out data frame.}

\item{type}{Backup in case can't figure out type.}
}
\value{
Either quick.table or invisible()
}
\description{
quick.reg method for class "manova".
}
\details{
Need to put some description of method here.
}
\examples{
quick.reg.manova(my.manova)
}
\keyword{Explore}
