% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tables.R
\name{quick.reg}
\alias{quick.reg}
\title{Regression Tables in Pixiedust}
\usage{
quick.reg(my.model, myDF = my.found.df, my.factor = NULL, SS.type = 3,
  pix.int = T, pix.method = "html", type = my.reg.type,
  test.stat = "Wilks")
}
\arguments{
\item{my.model}{Model. NOTE: If have factors, please place them first in the regression.}

\item{myDF}{Dataframe}

\item{my.factor}{If there are any factors, list them here.}

\item{SS.type}{Type of sums of squares to report (default = 3)}

\item{pix.int}{Should this be viewable or is this for a document/dashboard? (default=T)}

\item{pix.method}{Print method. (default="html")}

\item{type}{Type of regression? Currently supported: lm, glm (binary), manova}
}
\value{
Either pixiedust object or code (in HTML or LaTeX) for table
}
\description{
Beautiful tables adding sums of squares
and p-value formatting, then giving html or
latex output. If interactive and HTML, will show up
in viewer.
}
\examples{
quick.reg(my.model, myDF)
}
\keyword{Explore}
