% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.ROC.diag}
\alias{quick.ROC.diag}
\title{ROC Diagnositics table}
\usage{
quick.ROC.diag(my.glm, vals)
}
\arguments{
\item{my.glm}{Object of class "glm"}

\item{vals}{List of cutoff values to check.}
}
\value{
Matrix of values for each cutoff value.
}
\description{
Looks at Sensitivity, Specificity, and Area Under the Curve
for a range of cutoff values.
}
\examples{
quick.ROC.diag(my.glm,seq(0.4,0.6,by=0.05))
}
