% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.ROC}
\alias{quick.ROC}
\title{Quick ROC curve and Area Under ROC Curve}
\usage{
quick.ROC(my.glm, title = "ROC Curve", cutoff = 0.5, do.plot = T)
}
\arguments{
\item{my.glm}{Object of class "glm"}

\item{title}{Main title for graph (default: "ROC Curve")}

\item{cutoff}{Cutoff point? (default: 0.5)}

\item{do.plot}{Plot ROC Curve? (default: T)}
}
\value{
List with area under curve and plot
}
\description{
Takes GLM and computes ROC curve using ROCR library.
}
