% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.SAS.labels}
\alias{quick.SAS.labels}
\title{Quick SAS Factor Labels}
\usage{
quick.SAS.labels(my.df)
}
\arguments{
\item{my.df}{Data frame to get the new information}
}
\value{
Dataframe with label information
}
\description{
Why should you have to type them? A dialog will come up to ask you for the file.
}
