% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.search}
\alias{quick.search}
\title{Partial Name Search of Columns}
\usage{
quick.search(myDF, partial)
}
\arguments{
\item{myDF}{Dataframe}

\item{partial}{partial search as string}
}
\description{
Helper function to quickly get list of column names
based on partial piece.
}
\examples{
quick.search(myDF,"CAS")
}
\keyword{Explore}
