% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.table.R
\name{quick.table.check}
\alias{quick.table.check}
\title{Table Check}
\usage{
quick.table.check(q.tab)
}
\arguments{
\item{q.tab}{quick.table}
}
\value{
Logical of whether all matches
}
\description{
Checks to make sure that table and html table are the same before update
}
\keyword{internal}
