% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.table.multi.model}
\alias{quick.table.multi.model}
\title{Create multinomial omnibus tables}
\usage{
quick.table.multi.model(...)
}
\arguments{
\item{...}{List of models to be combined. One must have null model.}
}
\value{
Omnibus table
}
\description{
Create multinomial omnibus tables
}
