% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.table.R
\name{quick.table}
\alias{quick.table}
\title{Table Printer}
\usage{
quick.table(my.table, type, test.stat = "Pillai", print.type = "full",
  the.caption = NA, the.footer = NA, abbrev.length = ab.len,
  marginality = T, new.rownames.int = NULL, new.rownames.treat = NULL,
  swap.na = NULL, round.num = 2, col.names = my.colnames, print.now = T,
  show.footer = T, make.red = NULL, make.black = NULL, VIF = F,
  part.eta = F)
}
\arguments{
\item{my.table}{Final table ready to be turned into html}

\item{type}{Type (lm, glm, manova, ordinal)}

\item{test.stat}{Test stat used in MANOVA}

\item{print.type}{"full" for full html, "part" for just <div>}

\item{the.caption}{Caption for top of table}

\item{the.footer}{What is in the footer? This is a single, centered div tag.}
}
\value{
List of lists with 3 pieces: null model, pre models, and full models
}
\description{
Helper function to turn tables into html tables
}
\keyword{Explore}
