% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{quick.tasks-package}
\alias{quick.tasks}
\alias{quick.tasks-package}
\title{quick.tasks: Quick implementation of tables, figures, and other helper functions for quick analysis.}
\description{
Includes basic visualization implementations for tables from regression analyses.
Using PixieDust, LaTeX or HTML tables are output. Multiple summaries and results are included
in tables to more reflect what an analyst would need. Several extended options for tables are
also included (VIM, partial eta, glance stats, etc.). In addition, several helper functions for
tasks such as saving plots and seeing labels are included.
}
\details{
I will need to add a good set of details here.
}
\author{
\strong{Maintainer}: Christopher Kraner \email{ckraner19@gmail.com}

}
\keyword{internal}
