% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.to.na}
\alias{quick.to.na}
\title{Change certain numbers to NA}
\usage{
quick.to.na(my.df, nums)
}
\arguments{
\item{my.df}{Dataframe}

\item{nums}{list of numeric numbers to turn to NA indiscriminantly}
}
\value{
Dataframe
}
\description{
When NA do not come through from the SPSS dataset, apply value to all.
}
