% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.type}
\alias{quick.type}
\title{Get model type}
\usage{
quick.type(model)
}
\arguments{
\item{variable}{Variable from data set}
}
\value{
String of label
}
\description{
Helper function to see SPSS label
}
\examples{
quick.lavaan(myfit)
}
\keyword{Explore}
