% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.helpers.R
\name{quick.VIM}
\alias{quick.VIM}
\title{Quick VIM Aggregate Missing Plot}
\usage{
quick.VIM(my.df)
}
\arguments{
\item{my.df}{Dataframe}
}
\value{
VIM Object
}
\description{
Simple code, but used often enough that needed to be put in a function.
See the missing by calling the object. See graph with plot().
}
