% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.table.R
\name{update.quick.table}
\alias{update.quick.table}
\title{Table Update}
\usage{
\method{update}{quick.table}(q.tab, make.red = found.red,
  make.black = found.black, the.caption = my.caption, show.footer = T,
  new.rownames.int = NULL, new.rownames.treat = NULL, swap.na = NULL,
  the.round = my.round, print.type = "full", print.now = T,
  do.return = T)
}
\arguments{
\item{q.tab}{quick.table}
}
\value{
Logical of whether all matches
}
\description{
Update unchanged table
}
\keyword{Explore}
