% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.tasks.R
\name{quick.MICE}
\alias{quick.MICE}
\title{NOT NEEDED}
\usage{
quick.MICE(imputed.df, my.new.df)
}
\arguments{
\item{imputed.df}{Mids object from MICE}

\item{my.new.df}{Data frame that has been through label.explor.r}
}
\value{
Data frame with imputed values averaged and replaced.
}
\description{
Has some good code that can be reused later. Not needed atm.
}
